requirements.txt
flask==2.3.3
flask-SQLAlchemy==3.0.5
flask-JWT-Extended==4.5.3
flask-CORS==4.0.0
flask-Mail==0.9.1
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.1
python-multipart==0.0.6
psycopg2-binary==2.9.7
cloudinary==1.35.0 
werkzeug==2.3.7     
requests==2.31.0 
flask_jwt_extended
flask_cors
flask_mail
flask_sqlalchemy
psycopg2
cloudinary
Pillow